requirements.txt
pytest
pathspec
pydantic
pyyaml
datasets
dotenv
tiktoken
litellm
python-dotenv
langchain
langchain-community
langchain-huggingface
smolagents
unidiff
ruff
rich